Always draw background
The code used to avoid drawing background relied on clipping canvas with REPLACE region operation which no longer works as of API 28. Because of that I decided to get rid of this behavior and instead always draw the background.

WARNING: This is a breaking change which might require users of the library to give container views custom backgrounds in their code to avoid transparency issues.

Closes #9, #12